% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_covariate_matrix.R
\name{buildCovariateMatrix}
\alias{buildCovariateMatrix}
\title{Summarizes and standardizes covariates into a matrix}
\usage{
buildCovariateMatrix(modFrm, covariateFunctions)
}
\arguments{
\item{modFrm}{model frame object}

\item{covariateFunctions}{list of functions for summarizing covariates}
}
\description{
Summarizes and standardizes covariates into a matrix
}
