% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint_funcsynth.R
\name{confint.funcSynth}
\alias{confint.funcSynth}
\title{Confidence interval method for funcSynth objects}
\usage{
\method{confint}{funcSynth}(funcSynthObj, confidence = 0.95, type = c("syntheticControl", "gap"))
}
\arguments{
\item{funcSynthObj}{a functional synthetic control object}

\item{confidence}{level of confidence for the interval}

\item{type}{one (or both) of "syntheticControl", "gap". The former produces
a confidence interval for the snythetic control estimate and
the latter produces a confidence interval for the effect
estimate.}
}
\description{
Confidence interval method for funcSynth objects
}
