% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_synth.R
\name{funcSynth}
\alias{funcSynth}
\title{Functional Synthetic Control Method}
\usage{
funcSynth(
  formula,
  data,
  covariateFunctions = "mean",
  cfArgs = NULL,
  fpcaOptions = list(pre = list(), post = list()),
  V = c("FVE", "minMSE"),
  y = NULL,
  time = NULL,
  unit = NULL,
  intervention = NULL,
  treated = NULL,
  x = NULL,
  ...
)
}
\arguments{
\item{formula}{formula}

\item{data}{data}

\item{covariateFunctions}{named list of function names for summarizing
covariates. List names should correspond to
variable names.  Alternatively, instead of a list,
a single function may be supplied to be used on
all covariates.}

\item{cfArgs}{arguments for the covariate summary function}

\item{fpcaOptions}{options passed to FPCA()}

\item{V}{Either a numeric vector or matrix, or one of "minMSE" for
automated selection of V or "FVE" for percent variation explained.}

\item{y}{outcome}

\item{time}{time}

\item{unit}{unit}

\item{intervention}{indicator for intervention}

\item{treated}{indicator for treated}

\item{x}{covariates}

\item{...}{additional arguments passed to ipop() and optim().}
}
\value{
funkySynth object
}
\description{
Functional Synthetic Control Method
}
\details{
TODO
}
