% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lintify_data.R
\name{lintifyData}
\alias{lintifyData}
\title{Function subsets data to a single list unit or "lint".}
\usage{
lintifyData(u, y, t, ut, it, prepost, yt)
}
\arguments{
\item{u}{unit to subset}

\item{y}{outcome variable}

\item{t}{time variable}

\item{ut}{unit variable}

\item{it}{intervention variable}

\item{prepost}{pre or post intervention}

\item{yt}{selects where to make a list unit for outcome or time}
}
\description{
Function subsets data to a single list unit or "lint".
}
