% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listify_data.R
\name{listifyData}
\alias{listifyData}
\title{Converts modular synth object into list format needed for FPCA()}
\usage{
listifyData(mSynth)
}
\arguments{
\item{mSynth}{a modular synth object}
}
\description{
Converts modular synth object into list format needed for FPCA()
}
