% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_V.R
\name{lossV}
\alias{lossV}
\title{Loss function for optimizing V matrix}
\usage{
lossV(V, X0, X1, Z0, Z1, margin.ipop = 5e-04, sigf.ipop = 5, bound.ipop = 10)
}
\arguments{
\item{V}{V matrix for determining weights}

\item{X0}{FPCA score and covariate matrix for controls}

\item{X1}{FPCA score and covariate matrix for the treatment}

\item{Z0}{Outcome matrix for the controls}

\item{Z1}{Outcome matrix for the treatment}

\item{margin.ipop}{how close we get to the constraints; passed to ipop()}

\item{sigf.ipop}{Precision; passed to ipop()}

\item{bound.ipop}{Clipping bound for the variables; passed to ipop()}
}
\description{
Loss function for optimizing V matrix
}
