% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_dots.R
\name{matchDots}
\alias{matchDots}
\title{Internal function to handle the ... argument of funcSynth() appropariately.}
\usage{
matchDots(dots)
}
\arguments{
\item{dots}{additional arguments for optim() and ipop() from funcSynth call}
}
\description{
Internal function to handle the ... argument of funcSynth() appropariately.
}
