% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modular_synth.R
\name{modularSynth}
\alias{modularSynth}
\title{Creates modularSynth object for use in funcSynth()}
\usage{
modularSynth(y, time, unit, intervention, treated)
}
\arguments{
\item{y}{Numeric vector of outcome values.}

\item{time}{Either a numeric vector or vector of class Date, posixlt, or posixct}

\item{unit}{Character, numeric, or factor variable indicating unit.}

\item{intervention}{Binary or logical vector indicating whether an i
intervention has occurred on any unit at the corresponding
time.}

\item{treated}{Binary or logical variable indicating if a unit is ever
treated at any time.}
}
\description{
Creates modularSynth object for use in funcSynth()
}
