% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_weights.R
\name{optimizeWeights}
\alias{optimizeWeights}
\title{Find SCM weights using quadratic optimization}
\usage{
optimizeWeights(X0, X1, V, margin.ipop = 5e-04, sigf.ipop = 5, bound.ipop = 10)
}
\arguments{
\item{X0}{FPCA score and covariate matrix for controls}

\item{X1}{FPCA score and covariate matrix for the treatment}

\item{V}{Matrix specificing importance of FPCA scores and covariates for
determining weights.}

\item{margin.ipop}{how close we get to the constraints; passed to ipop()}

\item{sigf.ipop}{Precision; passed to ipop()}

\item{bound.ipop}{Clipping bound for the variables; passed to ipop()}
}
\description{
Find SCM weights using quadratic optimization
}
