% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcSynth.R
\name{plot.funcSynth}
\alias{plot.funcSynth}
\title{Generic plot method for funcSynth objects}
\usage{
\method{plot}{funcSynth}(x, ...)
}
\arguments{
\item{x}{a funcSynth object}

\item{...}{additional arguments passed to ggplot()}
}
\description{
Generic plot method for funcSynth objects
}
