% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_for_weights.R
\name{solveForWeights}
\alias{solveForWeights}
\title{Function that returns weights and determines V matrix if it is not supplied}
\usage{
solveForWeights(
  X0,
  X1,
  Z0 = NULL,
  Z1 = NULL,
  V,
  lambda = NULL,
  margin.ipop = 5e-04,
  sigf.ipop = 5,
  bound.ipop = 10,
  optimMethod = c("Nelder-Mead", "BFGS", "CG", "L-BFGS-B", "SANN", "Brent")
)
}
\arguments{
\item{X0}{FPCA score and covariate matrix for controls}

\item{X1}{FPCA score and covariate matrix for the treatment}

\item{Z0}{Outcome matrix for the controls}

\item{Z1}{Outcome matrix for the treatment}

\item{V}{Matrix specificing importance of FPCA scores and covariates for
determining weights.}

\item{lambda}{numeric vector of variances from FPCA}

\item{margin.ipop}{how close we get to the constraints; passed to ipop()}

\item{sigf.ipop}{Precision; passed to ipop()}

\item{bound.ipop}{Clipping bound for the variables; passed to ipop()}

\item{optimMethod}{One of "Nelder-Mead", "BFGS", "CG", "L-BFGS-B", "SANN",
"Brent"; passed to optim().  See optim() for more
information.}
}
\description{
Function that returns weights and determines V matrix if it is not supplied
}
