% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint_funcsynth.R
\name{confint.funcSynth}
\alias{confint.funcSynth}
\title{Confidence interval method for funcSynth objects}
\usage{
\method{confint}{funcSynth}(object, parm = c("syntheticControl", "gap"), level = 0.95, ...)
}
\arguments{
\item{object}{a functional synthetic control object}

\item{parm}{one (or both) of "syntheticControl", "gap". The former produces
a confidence interval for the snythetic control estimate and
the latter produces a confidence interval for the effect
estimate.}

\item{level}{level of confidence for the interval}

\item{...}{not used}
}
\description{
Confidence interval method for funcSynth objects
}
